pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Service with user endpoints"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "firebase-admin>=6.6.0",
    "pydantic-settings>=2.7.1",
    "uvicorn>=0.34.0",
]
